examples/layout/pyproject.toml
[tool.poetry]
name = "layout"
version = "0.1.0"
description = ""
authors = ["nebaskov <nikolai@niuitmo.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
